+++ src/main/java/nextstep/member/auth/Auth.java
@@ -1,4 +1,4 @@
-package nextstep.member;
+package nextstep.member.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
