+++ src/main/java/nextstep/member/auth/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package nextstep.member.application;
+package nextstep.member.auth;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
